feat(client): share native presentation and adapt contacts
