+++ apps/web/src/client-communication/index.css
@@ -39,3 +39,13 @@ body,
   height: 100%;
   background: var(--wk-bg-base, #fff);
 }
+
+.communication-contacts {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  min-height: 0;
+}
+
+.communication-contacts-header { display: none; }
+.communication-contacts-body { flex: 1; min-height: 0; }
