+++ apps/web/src/client-feature/desktop/presentation.css
@@ -1,0 +1,47 @@
+#root[data-desktop-platform] {
+  color: var(--wk-text-primary);
+  letter-spacing: 0;
+  --desktop-drag-region: drag;
+  -webkit-app-region: no-drag;
+}
+
+#root[data-desktop-drag-suspended="true"],
+#root[data-desktop-fullscreen="true"] {
+  --desktop-drag-region: no-drag;
+}
+
+#root[data-desktop-platform="darwin"] {
+  font-family: -apple-system, BlinkMacSystemFont, sans-serif;
+}
+
+#root[data-desktop-platform="win32"] {
+  font-family: "Segoe UI Variable", "Segoe UI", "Microsoft YaHei UI", sans-serif;
+}
+
+#root[data-desktop-platform] [data-desktop-header] {
+  box-sizing: border-box;
+  height: var(--desktop-header-height);
+  min-height: var(--desktop-header-height);
+  padding-left: calc(var(--desktop-safe-left) + var(--wk-sp-3));
+  padding-right: calc(var(--desktop-safe-right) + var(--wk-sp-3));
+  padding-top: 0;
+  padding-bottom: 0;
+  user-select: none;
+  -webkit-app-region: var(--desktop-drag-region, no-drag);
+  background: var(--desktop-header-background, var(--desktop-chrome-background, var(--wk-bg-surface)));
+  color: var(--desktop-chrome-foreground, var(--wk-text-primary));
+  border-bottom: 0;
+}
+
+/* Business descendants, including legacy div controls, stay interactive. */
+#root[data-desktop-platform] [data-desktop-header] * {
+  -webkit-app-region: no-drag;
+}
+
+#root[data-desktop-platform] [data-desktop-header] [data-desktop-chrome="layout"] {
+  -webkit-app-region: var(--desktop-drag-region, no-drag);
+}
+
+#root[data-desktop-focused="false"] [data-desktop-header] {
+  color: var(--wk-text-secondary);
+}
